pyproject.toml
[project]
name = "PaddleOCR-VL-BaiduAIStudio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]
